chore: moved device_filter to filip.utils.filter
BREAKING CHANGE: moved `filter_device_list` function to different module

For #147
