+++ filip/clients/ngsi_v2/iota.py
@@ -9,7 +9,7 @@ from filip.config import settings
 from filip.clients.base_http_client import BaseHttpClient
 from filip.models.base import FiwareHeader
 from filip.models.ngsi_v2.iot import Device, ServiceGroup
-from filip.utils.iot import filter_device_list
+from filip.utils.filter import filter_device_list
 
 
 class IoTAClient(BaseHttpClient):
